nftport/requirements.txt
aiohttp==3.8.1
Django==3.2.9
djangorestframework==3.12.4
ipdb==0.13.9
ipython==7.30.1
Pillow==8.4.0
psycopg2==2.9.2
requests==2.26.0

